Move CpuDeviceInterface to .cpp files where they belong
